requirements.txt
antlr4-python3-runtime==4.8
argon2-cffi==21.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cffi==1.15.0
click==8.0.3
cssselect2==0.4.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
imageio==2.9.0
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaitaistruct==0.9
kiwisolver==1.3.1
llvmlite==0.36.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.4
numba==0.53.1
numpy==1.19.5
numpy-vox-io==0.2
omegaconf==2.1.1
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
perlin-noise==1.8
perlin-numpy @ git+https://github.com/pvigier/perlin-numpy@6f077f811f5708e504732c26dee8f2015b95da0c
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pypng==0.0.21
pyrsistent==0.18.0
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
reportlab==3.6.2
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
svglib==1.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2020.9.3
tinycss2==1.1.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.6.0
